refactor: standardize imports using absolute paths from project root
- Update DataBackupTab component imports to use absolute paths
- Update SortingPrioritySettings imports to use absolute paths
- Consolidate BasicSettings page imports to use absolute paths
- Update GeneralTab component imports to use absolute paths
- Update PermissionOnboardingDialog imports to use absolute paths
- Move "dangerous-zone" tab from managedSite to general tab
